backend/requirements.txt
fastapi
uvicorn[standard]
PyGithub
gitpython
chromadb
httpx
python-dotenv
pydantic
pydantic-settings
